+++ CHANGES.rst
@@ -23,6 +23,7 @@ Bugfixes
 - Do not allow sending registration invitation reminders without the invitation
   link placeholder (:pr:`7093`)
 - Correctly log the user sending a registration invitation reminder (:pr:`7093`)
+- Fix error in weekday recurrence picker when using the Turkish locale (:pr:`7113`)
 
 Accessibility
 ^^^^^^^^^^^^^